Be smarter files removal between revision export
Since we now use export instead of checkout (to avoid keyword
expansion), the removed files will still be in the tree if not removed
by us.

We are using the changed paths to remove ourselves those that need to be
destroyed.

A better approach would be to retrieved the diff and apply it to the
previous revision tree.