Revert "Force end-of-line conversion behavior when needed"
This reverts commit 2b5badaa6af109e5b998929b6cba2ed1fa4a1566.

Strangely enough, in integration tests with a similar context, the
initial code works without needing to modify the svn patch...

Also, at runtime for the asf mirror loading, an assertion is raised
about size not matching after applying the patch and modifying
it. Indeed, we modify the patch to remove some characters (\r).

To conclude, this is not the right way, we revert.

If we want to continue on that path, we need to modify the file's
content after the patch has been applied.

Related T570