Revert "Patch file's content with eol conversion if needed"
This reverts commit 3ef381e378d883474643d091a9cb6db0a835dcee.

Trying to apply the modification after the patch application does not
work either.

Indeed, as explained in https://forge.softwareheritage.org/T570#10528,
the diffs received in this replay context are static. We are not dealing
with a working copy here. So, modifying a file outside of the patch and
trying to apply another future static patch on the same file in the
future won't work. The diff won't be applicable correctly. Thus, highly
likely the length won't match thus raising the error mentioned in
e61e4c4335ef173225a94ef7ee635672621be21a.

Related T570