ra.py: Normalize line endings when svn:eol-style property is set
This commit ensures that the end of line style for each file extracted
from a svn revision respects the one defined by the svn:eol-style
property if defined.

Indeed, it exists some corner cases where badly formatted content
data were saved into a subversion repository (for instance a file
with CRLF line endings but with the property svn:eol-style set
to native) that prevent the subversion loader to succeed.

Closes T570