svn/loader: Add efficient incremental loader based on partial dumps
This commits adds a new loader class SvnLoaderFromRemoteDump enabling
to load svn repositories in an incremental way. The loader is based
on the creation of dump files generated with the rsvndump tool. Using
dump files allow to greatly speedup the loading process compared
to a client/server based approach.

Related T1161