svn.loader: Add efficient loader based on remote dumps
This commits adds a new loader class SvnLoaderFromRemoteDump enabling
to load svn repositories in a more efficient way. The loader is based
on the creation of dump files generated with the svnrdump tool. Using
dump files allow to greatly speedup the loading process compared
to a client/server based approach.

Closes T1161