loader: Fix SvnLoaderFromDumpArchive and add test
Previously SvnLoaderFromDumpArchive was trying to connect to the origin
URL provided as parameter while it should not happen.

For the dump archive case, the subversion repository is recreated locally
from a dump file so the subversion repository URL must correspond to
the local path of that repository.

Also add missing test for that specific loader.

Closes T3686