ra: Do not process end of lines for svn links
In a subversion repository, files with the svn:special property set
correspond to svn links and will be exported as symbolic links when
doing a checkout or export operation.

A subversion link is stored as a special file in the repository with
the following content: "link <path_to_target>".

So when the svn:eol-style property is also set on a svn link, we must
ensure that its content will not be overriden by the content of the
link target with processed end of lines. Otherwise the reconstructed
filesystem for a revision will differ from the one resulting from
an export operation on that revision.

Related to T3695