ra: Handle unsetting svn:special property on svn links
Subversion links have the svn:special property set when added in
a repository.

Nevertheless, it exist cases when such a property gets unset on
svn links in another revision. In that case, the svn links will
no longer be extracted as symbolic links during a checkout or
export operation but rather to regular files with the following
content: "link <path_to_target>".

So when the svn:special property gets unset on a svn link, ensure
to remove the symbolic link previously created by the ra module
and replace it by a regular file containing the svn link info.
Otherwise the reconstructed filesystem for a revision will differ
from the one resulting from an export operation on that revision.

Related to T3695