ra: Fix export of non link file with svn:special property set
When a file with the svn:special property set is not a svn link,
the svn export operation will extract a truncated version of that
file if it contains a null byte (see create_special_file_from_stream
implementation in libsvn_subr/subst.c).

So ensure to produce the same file as the svn export operation
otherwise the reconstructed file system will differ.

Related to T3695