loader: Fix error when first revision number is not one
It exist cases where the first revision number returned by a subversion
server has a value greater than one.

The loader did not expect such cases in its current implementation and
was raising an error preventing the loading of such repository in the
archive (apart when using the SvnLoaderFromRemoteDump class which
always dumps revisions starting from number one).

So ensure to cover that case in loader implementation.

Related to T3701