loader: Implement post_load to check for revision divergence
When all revisions of a subversion repository have been processed and
loaded, check if the reconstructed filesystem for the last revision
is consistent with the one obtained by a svn export operation.

If it is not, an exception will be raised in order to report the
issue and mark the visit as partial.

Previously that kind of issue could only be detected by triggering
a new load of the repository.

Related to T3695