SvnLoaderFromRemoteDump: Drop dump when svn repository is mounted
This will decrease the disk pressure currently existing when ingesting a svn repository
out of a remote dump. We first fetch a dump, then mount a svn repository out of it, and
at last we ingest the repository and have a growing copy on disk [1]. So we are up to 3
copies which takes lots of disk space.

With the following commit, we take down the 1 unnecessary copy (svn dump) as soon as
possible (when the svn repository is mounted which we ingest).

[1] implementation detail for speed, we iterate over the commit log, apply deltas (and
computing swh hashes) along the way on disk.

Related to T3719