loader: Optimize SvnLoaderFromRemoteDump use on stale repository
When trying to load multiple times a stale subversion repository with
SvnLoaderFromRemoteDump, first check if the last loaded revision in the
archive is different from the one on the remote subversion server.

If they are identical skip the dump, mount and load phases in order
to gain some disk space and processing time on the celery worker
executing the loading task.

Related to T3719