svn: Ensure proper incremental loading of a repository
When performing an incremental loading of a subversion repository, i.e. only
load new revisions issued since the last loading, we need to replay the whole
set of path modifications since the first revision in order to restore possible
file states induced by setting svn properties on those files.

For instance if a file got the svn:eol-style property set in a revision lesser
than the last one loaded into the archive, we will miss that information if
we do not start replaying the paths modifications since the first revision.

So ensure to replay path modifications since first revision and start yielding
new data to archive once we reached the revision to resume the loading from.