svn: Use a temporary directory when checking if history got altered
Previously when checking if svn history got altered when performing
incremental load, the last loaded subversion revision was exported
in the same directory than the one used to replay the svn revisions.

As this directory was not cleared between the history check and
the revisions replay, this could cause errors when replaying revisions
and thus make the loading fails.

So ensure to use a new temporary directory when checking altered
history and clear it when that process has been performed.

Also remove now useless reset of ra.Replay object.

Related to T3747