ra: Persist path executable flag in FileState when replaying revisions
When a file is marked as executable through the use of the svn:executable
property, that info must be persisted across replayed revisions.

Indeed, such a file can also be a svn link but when it is turned back to a
regular file (by unsetting svn:special property and modifying its content)
the execution rights should also be set on it as with a subversion export.

Without persisting the executable info, the execution rights will not be
restored when such a case appears.

Related to T3695