ra: Persist link flag in FileState when replaying revisions
When adding a svn link that will be exported as a symbolic link,
that information must be persisted across revisions replay.

Previously if a link was modified in another revision, like for
instance adding svn:eol-style property, this could lead to
removing the exported symbolic link while it should not.

Related to T3695