ra: Fix parsing of svn link with multiple lines
Some svn links might contain multiple lines if a user added an
end of line character in the first line.

When exporting a link, subversion will only consider the first
line content to create a symbolic link so we need to apply the
same processing or the link we create will differ from the
exported one.

Related to T3695