ra: Rework export of non link file with svn:special property set
The fix introduced in 58a07c67d590 to ensure reconstruction of non link
binary file with svn:special property set will be the same as an export
operation was not generic enough to handle all possible cases that can
be found in the wild.

So rework it by explicitely exporting that specific file with a subversion
client when it is needed.

Related to T3695