loader: Do not raise error when reloading a repo not fully ingested
When a subversion repository did not get fully ingested during a first
load (an error occurred when processing a revision that stopped the
loading), trying to reloading it again will raise a ValueError instead
of generating the same snapshot as the previous visit.

So ensure to initialize the _snapshot and _last_revision class member
of the loader to the latest loaded snapshot and revision if they exist
to avoid raising an error and generate an uneventful visit.