ra: Add support for subversion external definitions
Subversion external definitions set on directories through the use of the
svn:externals property are now handled by the loader.

As with a svn export operation, externals will be attempted to be exported
in the paths they are defined. If an external is no longer valid (404),
the error will be ignored and the next one will be processed.

The implementation takes care of keeping the reconstructed repository
filesystem for a revision in sync with a svn export operation while
externals are added, updated or removed across revisions replay.

Related to T611