ra: Put externals in cache to avoid exporting them again
Some subversion repositories can define same external on different paths.

In order to avoid exporting it multiple times, which consumes network bandwith
and slows down the loading, save the exported external in a temporary directory
on the local filesystem and reuse that copy when the external is set on a path.

Also ensure all the temporary directories created for externals will be deleted
at the end of the loading process.

Related to T611