replay: Use peg_rev parameter of client.export instead of rev
It exist cases where there is an ambiguity for subversion to export a path
from a given revision.

See https://edoras.sdsu.edu/doc/svn-book-html-chunk/svn.advanced.pegrevs.html
for more details.

To remove the ambiguity, the peg_rev parameter of the export operation must
be used instead of the rev one.

Related to T3870