replay: Fix detection of external paths to remove
When trying to detect if some external paths previously set in a revision
need to be removed when the content of a svn:externals property is modified,
the external URL and the target revision must also be taken into account.

Previously, only the external path was considered but the external URL
or the revision might have changed. In that case, the external path must
also be removed before trying to export it again, otherwise the loader
will consider it already exported and skip the operation.

Related to T3870