replay: Handle recursive external definition
It is possible that an external definition lead to a recursive subversion
export operation (https://issues.apache.org/jira/browse/SVN-1703).

As that recursive export will crash the loader and make the loading failed,
the only solution for that edge case is to stop processing externals while
a recursive external is detected in a revision.

We will then ignore externals when exporting a revision to check for divergence
with the recontructed filesystem.

Related to T3870