replay: Prevent external path removal when defined on multiple dirs
The same absolute external path can be defined on multiple directories
in the repository.

To prevent removal of an external path while it is still defined in
another directory, we need to track the number of references of the
absolute external path by incrementing / decrementing a counter
when adding / removing an external.

Related to T611