replay: Fix external tuple use in process_external method
Some required changes were missed in 20c1445fb11e.

The external tuple created at the beginning of the process_external method
was missing a third boolean member.

The externals defintions in DirState was also missing that boolean member.

Test checking if an external was already exported was invalid.

Also alias the type of that tuple for better code readability.

Related to T611