loader: Simplify the handing of parent revision
In the current loader implementation, a revision different from the
first one has a single parent revision corresponding to the previously
processed one.

In order to simplify the handling of the parent revision, remove
the use of the revision_parents dict and simply store the previously
processed revision id in a parents tuple variable while iterating on
the revisions log. In case of incremental loading, that tuple will be
initialized from the latest revision loaded into the archive.

This change is required to allow the loading of svn subprojects.

Related to T3896