loader: Fix repo sub-tree loading when using dump loaders
When dumping a repository sub-tree using svnrdump, svnrdump filters the
repository paths outside of the sub-tree but still dumps all commits of
the root repository. This means that the produced dump might contain
empty commits if those modify paths outside of the sub-tree.

So ensure to have the same loading behavior as with SvnLoader class,
that communicates directly with the remote repository, when loading
a repository sub-tree from a remote dump. These changes ensure no
empty commits will be archived and correct root directory when loading
a sub-tree with SvnLoaderFromDumpArchive or SvnLoaderFromRemoteDump.

Related to T3896