replay: Fix handling of directory symlink in external
As we use distutils.dir_util.copy_tree to copy an external exported
in a cache directory to the reconstructed filesystem of the subversion
repository, we need to ensure sub-directories of the external will
exist in the destination path to avoid copy errors.

However, if a sub-directory is a symlink, we must not create the
directory in the destination path as copy_tree will fail attempting
to recreate the symlink.