svn: Prefer to svn cleanup path instead of removing it before checkout
In production, we have errors reporting that path is already locked or that
the working copy cannot be locked when attempting to perform a checkout
operation.

For some reasons, it seems that attempting to remove the checkout path before
that operation is not sufficient for some cases.

So prefer to svn cleanup the working copy to ensure remaining locks will be
removed, checkout operation can then be safely retried.