utils: Ensure to return correct revision in parse_external_definition
An external definition can be of the following form (where XXX and YYY
are revision numbers):

	-r XXX <repo_url>@YYY

In that case, the official subversion client will export the revision
XXX of the external repository.

So ensure to have the same behavior when the subversion loader processes
a repository with such external defintion in it.