loader: Compress dump file and rework truncated dump handling
When dumping a subversion repository to file before loading it, compress
that file using gzip while producing it. It enables to save significant
disk space while dumping a large repository.

Also rework the way how truncated dump is handled now dump file is
compressed by providing the expected max revision number to be loaded
by svnadmin. If the number of loaded revisions matches, we can safely
continue the partial loading of the repository.