utils: Add support for parsing externals with revisions as dates
Subversion allow to specify the revision for an external definition as a
date instead of an integer identifier.

So when encountering such case, get the HEAD revision number for the
external at the specified date in order to export the correct version
of the files targeted by the external.

Related to #4727