replay: Filter externals to copy in copyfrom operations
When a directory is copied from another one in a previous revision,
externals must be copied only if they have been defined in a revision
greater or equal to the revision the directory is copied from.

So store the revision number an external is defined and use it to filter
externals when performing copyfrom operations.