Add svn directory loader in charge of ingesting specific svn tree
This overrides the default core class DirectoryLoader's fetch_data method to export an
svn tree at a specific commit or tag. The remaining behavior is the same as the
BaseDirectoryLoader. Checks the checksums of the tree is ok, then ingest the DAG objects
into the archives (including the NAR as extid).

Refs. swh/meta#4979