svn_repo: Optimize export_temporary performances significantly
The export_temporary method of the SvnRepo class exports the content of
a subversion repository at a given revision in a temporary directory.

As we also export the externals that might be associated to some paths
in the repository, we first need to get all the svn:externals property
values in order to determine if there is recursive or relative externals
and adjust some export parameters accordingly.

While that operation is fast when the subversion repository is hosted
locally, it is terribly slow when the repository is hosted on a remote
server. Indeed a recursive propget operation on a remote server sends
a lot of network requests which slows down quite a lot the process,
especially with large repositories.

To improve the performances, the previous implementation was doing a
full checkout of the repository to local filesystem and gets svn:externals
property values from it. Nevertheless, that process is time consuming for
large repositories and it can consume a lot of disk space.

In order to remove that bottleneck and improve overall performances for
getting all properties values, introduce a C++ extension module for Python
that implements a fast way to crawl all paths of a repository and their
associated properties. Unlike "svn ls --depth infinity" or "svn propget -R"
commands it performs only one SVN request over the network, hence saving
time especially with large repositories.
The code is freely inspired from the fast-svn-crawler project by Dmitry
Pavlenko (https://sourceforge.net/projects/fastsvncrawler/).

The obtained speedup is quite impressive, on a large remote repository
listing all paths using "svn ls --depth infinity" or gettings all svn:externals
property values using "svn propget -R" takes around one hour while it takes
only a couple of minutes using the approach implemented in the C++ extension
module. Using that approach also enables to save disk space as we no longer
need to perform a full checkout of the repository.

This change should greatly improve the performances when reloading a svn
repository already visited by Software Heritage. Indeed, before the possible
archiving of new commits issued since last visit, the loader checks that a
repository has not been altered by calling the export_temporary method using
the remote repository URL.