svn_repo: Optimize export of a remote subversion sub-path
Previously when exporting a sub-path of a remote subversion repository
over the network, the full repository was exported and the local path
targeting the sub-path was returned. This is no really optimal in terms
of network bandwidth if the repository filesystem is large but it was
implemented like this to ensure all tests related to sub-paths export
were passing regardless the subversion loader class used: either SvnLoader
or SvnLoaderFromRemoteDump.

After some analysis, it turned out that it was possible to avoid exporting
the full repository but only the request sub-path when using the SvnLoader
class. So modify the SvnRepo class to ensure that behavior and save some
network bandwdith when dealing with a large repository.

These changes in the SvnRepo class induce some in the replay module to ensure
all tests still pass and it also enables to remove a no longer needed optional
parameter to the class constructor.