svn_repo: Handle anonymous credentials in get_svn_repo
Some remote subversion repositories in the wild require to use anonymous
credentials to obtain a read access. Those credentials can be either
'anonymous/' or 'anonymous/anonymous' in most of the cases.

Previously, it was required to add the credentials in the origin URL using
basic authentication syntax. While such URL can be provided to the Save Code
Now service, it cannot when svn URLs are coming from a lister.

To workaround this, try to connect using these anonymous credentials in the
get_svn_repo function when a connection error happens.

This should also simplify the submission of Save Code Now requests when
submitting a subversion origin that requires anonymous credentials.