loader: Prevent errors when loading repository sub-path from a dump
When loading a sub-path of a repository from a dump file it has been
observed it is less error-prone to dump the whole repository as some
partial dumps fail to be loaded by svnadmin or svnrdump can end up
with error.

In that case, we try to produce a dump file for the whole repository
and we stop to dump revisions once the last one modifying the repository
sub-path was dumped (revisions not modifying sub-path are then filtered
out during the loading process).