loader: Use proper way to get latest revision where a path was modified
This fixes a regression introduced in f94f1d0.

Using the svn info command to get the latest revision where a path was
modified is not reliable as it does not handle the case where a path was
copied from a previous revision. In that case, the info command is returning
the latest revision where the path was modified before the copy. So the
loading of a repository subpath that was copied could not succeed as the
revision where the copy happened was not considered.

Instead use the svn log command that ensures the revision where the copy
of the path happened is returned.