loader, replay: Handle corrupted externals
Some externals can be corrupted and their export do not successfully finish
in that case. As the subversion client does not guarantee an order when
exporting files and directories of an external, ensure to remove corrupted
external paths from the reconstructed filesystem and in the temporary export
of a revision to avoid directory hashes mismatch when checking for integrity.