utils: Allow partial loading when a repo dump cannot be fully imported
Some repository dumps fail to be fully imported for various reasons: checksum
mismatch, file not found, ...

Previously the loader was failing when encountering such cases, however a
subset of revisions were still imported so it is possible to do a partial
load of such repositories.

Ensure to do a partial load of such repository by modifying the implementation
of the init_svn_repo_from_dump function.